WIP: use jest mocks in tests
