+++ .yarn/sdks/eslint/package.json
@@ -1,6 +1,6 @@
 {
   "name": "eslint",
-  "version": "8.14.0-sdk",
+  "version": "8.20.0-sdk",
   "main": "./lib/api.js",
   "type": "commonjs"
 }
