+++ .yarn/sdks/typescript/package.json
@@ -1,6 +1,6 @@
 {
   "name": "typescript",
-  "version": "4.6.3-sdk",
+  "version": "4.7.4-sdk",
   "main": "./lib/typescript.js",
   "type": "commonjs"
 }
