+++ packages/react-select-async-paginate/package.json
@@ -50,6 +50,7 @@
     "@storybook/addon-docs": "^6.5.9",
     "@storybook/react": "^6.5.9",
     "@types/jest": "^28.1.6",
+    "@types/node": "^18.7.9",
     "@types/react": "^18.0.15",
     "@types/react-test-renderer": "^18.0.0",
     "@typescript-eslint/eslint-plugin": "^5.30.7",
@@ -65,6 +66,7 @@
     "eslint-plugin-react": "^7.30.1",
     "eslint-plugin-react-hooks": "^4.6.0",
     "jest": "^28.1.3",
+    "jest-mock-console": "^2.0.0",
     "react": "^18.2.0",
     "react-select": "^5.4.0",
     "react-test-renderer": "^18.2.0",
