+++ packages/react-select-async-paginate/src/__tests__/getInitialCache-test.ts
@@ -1,0 +1,37 @@
+import type {
+  GroupBase,
+} from 'react-select';
+
+import { getInitialCache } from '../getInitialCache';
+
+import type {
+  UseAsyncPaginateBaseParams,
+} from '../types';
+
+const defaultParams: UseAsyncPaginateBaseParams<unknown, GroupBase<unknown>, unknown> = {
+  loadOptions: () => ({
+    options: [],
+  }),
+  inputValue: '',
+  menuIsOpen: false,
+};
+
+test('should return initial cache', () => {
+  const additional = Symbol('additional');
+
+  const params = {
+    ...defaultParams,
+    additional,
+    defautAdditional: {
+      page: 2,
+    },
+  };
+
+  expect(getInitialCache(params)).toEqual({
+    isFirstLoad: true,
+    options: [],
+    hasMore: true,
+    isLoading: false,
+    additional,
+  });
+});
