+++ packages/react-select-async-paginate/src/__tests__/useAsyncPaginateBase-test.ts
@@ -1,21 +1,23 @@
 import {
-  useCallback as reactUseCallbact,
-  useState as reactUseState,
-  useRef as reactUseRef,
+  useEffect,
+  useRef,
+  useState,
 } from 'react';
+
 import useIsMountedRef from 'use-is-mounted-ref';
-import sleepLib from 'sleep-promise';
+
+import type {
+  GroupBase,
+} from 'react-select';
 
 import { defaultShouldLoadMore } from '../defaultShouldLoadMore';
 import { defaultReduceOptions } from '../defaultReduceOptions';
+import { getInitialOptionsCache } from '../getInitialOptionsCache';
+import { requestOptions } from '../requestOptions';
 
 import {
   increaseStateId,
-  getInitialOptionsCache,
-  getInitialCache,
-  validateResponse,
-  requestOptions,
-  useAsyncPaginateBasePure,
+  useAsyncPaginateBase,
 } from '../useAsyncPaginateBase';
 
 import type {
@@ -24,12 +26,22 @@ import type {
   UseAsyncPaginateBaseParams,
 } from '../types';
 
+jest.mock('react', () => ({
+  ...jest.requireActual('react'),
+
+  useState: jest.fn()
+    .mockReturnValue([1, () => undefined]),
+
+  useEffect: jest.fn(),
+  useRef: jest.fn(),
+  useCallback: jest.fn(<T extends Function>(callback: T) => callback),
+}));
+
 jest.mock('use-is-mounted-ref');
+jest.mock('../getInitialOptionsCache');
+jest.mock('../requestOptions');
 
-type UseStateResult = [number, (nextCache: OptionsCache<any, any, any>) => void];
-type UseStateArgs = [() => number];
-
-const defaultCacheItem: OptionsCacheItem<any, any, any> = {
+const defaultCacheItem: OptionsCacheItem<unknown, GroupBase<unknown>, unknown> = {
   options: [],
   isLoading: false,
   isFirstLoad: false,
@@ -37,7 +49,7 @@ const defaultCacheItem: OptionsCacheItem<any, any, any> = {
   additional: null,
 };
 
-const defaultParams: UseAsyncPaginateBaseParams<any, any, any> = {
+const defaultParams: UseAsyncPaginateBaseParams<unknown, GroupBase<unknown>, unknown> = {
   loadOptions: () => ({
     options: [],
   }),
@@ -45,7 +57,30 @@ const defaultParams: UseAsyncPaginateBaseParams<any, any, any> = {
   menuIsOpen: false,
 };
 
-const makeUseRef = ({
+beforeEach(() => {
+  (useRef as jest.Mock)
+    .mockReturnValueOnce({
+      current: true,
+    })
+    .mockReturnValueOnce({
+      current: defaultParams,
+    })
+    .mockReturnValueOnce({
+      current: {},
+    });
+
+  (getInitialOptionsCache as jest.Mock).mockImplementation(
+    jest.requireActual('../getInitialOptionsCache').getInitialOptionsCache,
+  );
+
+  (requestOptions as jest.Mock).mockResolvedValue(undefined);
+});
+
+afterEach(() => {
+  jest.clearAllMocks();
+});
+
+const mockUseRef = ({
   isInit = {
     current: true,
   },
@@ -55,20 +90,26 @@ const makeUseRef = ({
   optionsCache = {
     current: {},
   },
-}): typeof reactUseRef => jest.fn()
-  .mockReturnValueOnce(isInit)
-  .mockReturnValueOnce(params)
-  .mockReturnValueOnce(optionsCache);
+}: {
+  isInit?: {
+    current: boolean;
+  };
 
-const defaultUseEffect = (): void => {};
+  params?: {
+    current: UseAsyncPaginateBaseParams<unknown, GroupBase<unknown>, unknown>;
+  };
 
-const defaultUseCallback: typeof reactUseCallbact = (fn) => fn;
+  optionsCache?: {
+    current: OptionsCache<unknown, GroupBase<unknown>, unknown> | null;
+  };
+}) => {
+  (useRef as jest.Mock).mockReset();
 
-const defaultValidateResponse = (): void => {};
-
-const defaultRequestOptions = async (): Promise<void> => {};
-
-const defaultUseState = (): [number, () => void] => [1, (): void => {}];
+  (useRef as jest.Mock)
+    .mockReturnValueOnce(isInit)
+    .mockReturnValueOnce(params)
+    .mockReturnValueOnce(optionsCache);
+};
 
 beforeEach(() => {
   (useIsMountedRef as jest.Mock).mockReturnValue({
@@ -86,226 +127,10 @@ describe('increaseStateId', () => {
   });
 });
 
-describe('getInitialOptionsCache', () => {
-  test('should return empty options cache', () => {
-    const initialOptionsCache = getInitialOptionsCache(defaultParams);
-
-    expect(initialOptionsCache).toEqual({});
-  });
-
-  test('should return options cache with "options" prop', () => {
-    const options = [
-      {
-        label: 'label 1',
-        value: 'value 1',
-      },
-      {
-        label: 'label 2',
-        value: 'value 2',
-      },
-    ];
-
-    const initialOptionsCache = getInitialOptionsCache({
-      ...defaultParams,
-      options,
-    });
-
-    expect(initialOptionsCache).toEqual({
-      '': {
-        isFirstLoad: false,
-        isLoading: false,
-        hasMore: true,
-        options,
-        additional: undefined,
-      },
-    });
-  });
-
-  test('should return options cache with "defaultOptions" prop', () => {
-    const options = [
-      {
-        label: 'label 1',
-        value: 'value 1',
-      },
-    ];
-
-    const defaultOptions = [
-      {
-        label: 'label 2',
-        value: 'value 2',
-      },
-      {
-        label: 'label 3',
-        value: 'value 3',
-      },
-    ];
-
-    const initialOptionsCache = getInitialOptionsCache({
-      ...defaultParams,
-      options,
-      defaultOptions,
-    });
-
-    expect(initialOptionsCache).toEqual({
-      '': {
-        isFirstLoad: false,
-        isLoading: false,
-        hasMore: true,
-        options: defaultOptions,
-        additional: undefined,
-      },
-    });
-  });
-
-  test('should set "additional" with "additional" param in initialOptionsCache', () => {
-    const options = [
-      {
-        label: 'label 1',
-        value: 'value 1',
-      },
-    ];
-
-    const defaultOptions = [
-      {
-        label: 'label 2',
-        value: 'value 2',
-      },
-      {
-        label: 'label 3',
-        value: 'value 3',
-      },
-    ];
-
-    const initialOptionsCache = getInitialOptionsCache({
-      ...defaultParams,
-      options,
-      defaultOptions,
-      additional: {
-        page: 1,
-      },
-    });
-
-    expect(initialOptionsCache).toEqual({
-      '': {
-        isFirstLoad: false,
-        isLoading: false,
-        hasMore: true,
-        options: defaultOptions,
-        additional: {
-          page: 1,
-        },
-      },
-    });
-  });
-
-  test('should set "additional" with "defaultAdditional" param in initialOptionsCache', () => {
-    const options = [
-      {
-        label: 'label 1',
-        value: 'value 1',
-      },
-    ];
-
-    const defaultOptions = [
-      {
-        label: 'label 2',
-        value: 'value 2',
-      },
-      {
-        label: 'label 3',
-        value: 'value 3',
-      },
-    ];
-
-    const initialOptionsCache = getInitialOptionsCache({
-      ...defaultParams,
-      options,
-      defaultOptions,
-      additional: {
-        page: 1,
-      },
-      defaultAdditional: {
-        page: 2,
-      },
-    });
-
-    expect(initialOptionsCache).toEqual({
-      '': {
-        isFirstLoad: false,
-        isLoading: false,
-        hasMore: true,
-        options: defaultOptions,
-        additional: {
-          page: 2,
-        },
-      },
-    });
-  });
-
-  test('should not set options cache if "defaultOptions" is true', () => {
-    const options = [
-      {
-        label: 'label 1',
-        value: 'value 1',
-      },
-    ];
-
-    const initialOptionsCache = getInitialOptionsCache({
-      ...defaultParams,
-      options,
-      defaultOptions: true,
-    });
-
-    expect(initialOptionsCache).toEqual({});
-  });
-});
-
-describe('getInitialCache', () => {
-  test('should return initial cache', () => {
-    const additional = Symbol('additional');
-
-    const params = {
-      ...defaultParams,
-      additional,
-      defautAdditional: {
-        page: 2,
-      },
-    };
-
-    expect(getInitialCache(params)).toEqual({
-      isFirstLoad: true,
-      options: [],
-      hasMore: true,
-      isLoading: false,
-      additional,
-    });
-  });
-});
-
-describe('validateResponse', () => {
-  const fakeConsole = {
-    error: (): void => {},
-  };
-
-  test('should throw error if response falsy', () => {
-    expect(() => {
-      validateResponse(fakeConsole as Console, null);
-    }).toThrowError();
-  });
-
-  test('should throw error if list of options is not array', () => {
-    expect(() => {
-      validateResponse(fakeConsole as Console, {
-        options: 123,
-      });
-    }).toThrowError();
-  });
-});
-
-describe('useAsyncPaginateBasePure', () => {
+describe('useAsyncPaginateBase', () => {
   test('should call getInitialOptionsCache on init', () => {
-    const getInitialOptionsCacheParam = jest.fn()
-      .mockReturnValue({});
+    (getInitialOptionsCache as jest.Mock).mockReset();
+    (getInitialOptionsCache as jest.Mock).mockReturnValue({});
 
     const options = [
       {
@@ -329,18 +154,13 @@ describe('useAsyncPaginateBasePure', () => {
       page: 2,
     };
 
-    useAsyncPaginateBasePure(
-      makeUseRef({
-        optionsCache: {
-          current: null,
-        },
-      }),
-      defaultUseState,
-      defaultUseEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCacheParam,
-      defaultRequestOptions,
+    mockUseRef({
+      optionsCache: {
+        current: null,
+      },
+    });
+
+    useAsyncPaginateBase(
       {
         ...defaultParams,
         options,
@@ -349,8 +169,8 @@ describe('useAsyncPaginateBasePure', () => {
       },
     );
 
-    expect(getInitialOptionsCacheParam).toHaveBeenCalledTimes(1);
-    expect(getInitialOptionsCacheParam.mock.calls[0][0]).toEqual({
+    expect(getInitialOptionsCache).toHaveBeenCalledTimes(1);
+    expect(getInitialOptionsCache).toHaveBeenCalledWith({
       ...defaultParams,
       options,
       defaultOptions,
@@ -360,70 +180,45 @@ describe('useAsyncPaginateBasePure', () => {
 
   test('should provide deps to first useEffect', async () => {
     const deps = [1, 2, 3];
-    const useEffect = jest.fn();
 
-    useAsyncPaginateBasePure(
-      makeUseRef({}),
-      (defaultUseState as unknown as typeof reactUseState),
-      useEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      defaultRequestOptions,
+    useAsyncPaginateBase(
       defaultParams,
       deps,
     );
 
-    expect(useEffect.mock.calls[0][1]).toBe(deps);
+    expect((useEffect as jest.Mock).mock.calls[0][1]).toBe(deps);
   });
 
   test('should not load options from first useEffect if "defaultOptions" is not true', async () => {
-    const useEffect = jest.fn();
-    const requestOptionsParam = jest.fn();
-
-    useAsyncPaginateBasePure(
-      makeUseRef({}),
-      (defaultUseState as unknown as typeof reactUseState),
-      useEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      requestOptionsParam,
+    useAsyncPaginateBase(
       {
         ...defaultParams,
         defaultOptions: [],
       },
     );
 
-    useEffect.mock.calls[0][0]();
+    (useEffect as jest.Mock).mock.calls[0][0]();
 
-    expect(requestOptionsParam).toHaveBeenCalledTimes(0);
+    expect(requestOptions).toHaveBeenCalledTimes(0);
   });
 
   test('should load options from first useEffect if "defaultOptions" is true', async () => {
-    const useEffect = jest.fn();
-    const requestOptionsParam = jest.fn();
-
-    useAsyncPaginateBasePure(
-      makeUseRef({}),
-      (defaultUseState as unknown as typeof reactUseState),
-      useEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      requestOptionsParam,
+    useAsyncPaginateBase(
       {
         ...defaultParams,
         defaultOptions: true,
       },
     );
 
-    useEffect.mock.calls[0][0]();
+    (useEffect as jest.Mock).mock.calls[0][0]();
 
-    expect(requestOptionsParam).toHaveBeenCalledTimes(1);
+    expect(requestOptions).toHaveBeenCalledTimes(1);
   });
 
   test('should not reset options cache from first useEffect on initial render', async () => {
+    const setStateId = jest.fn();
+    (useState as jest.Mock).mockReturnValue([1, setStateId]);
+
     const isInit = {
       current: true,
     };
@@ -434,26 +229,16 @@ describe('useAsyncPaginateBasePure', () => {
       },
     };
 
-    const setStateId = jest.fn();
-    const useEffect = jest.fn();
-    const useState = jest.fn<UseStateResult, UseStateArgs>()
-      .mockReturnValue([2, setStateId]);
+    mockUseRef({
+      isInit,
+      optionsCache,
+    });
 
-    useAsyncPaginateBasePure(
-      makeUseRef({
-        isInit,
-        optionsCache,
-      }),
-      (useState as unknown as typeof reactUseState),
-      useEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      defaultRequestOptions,
+    useAsyncPaginateBase(
       defaultParams,
     );
 
-    useEffect.mock.calls[0][0]();
+    (useEffect as jest.Mock).mock.calls[0][0]();
 
     expect(isInit.current).toBe(false);
     expect(optionsCache.current).toEqual({
@@ -463,6 +248,9 @@ describe('useAsyncPaginateBasePure', () => {
   });
 
   test('should reset options cache from first useEffect on not initial render', async () => {
+    const setStateId = jest.fn();
+    (useState as jest.Mock).mockReturnValue([1, setStateId]);
+
     const isInit = {
       current: false,
     };
@@ -479,69 +267,36 @@ describe('useAsyncPaginateBasePure', () => {
       },
     };
 
-    const setStateId = jest.fn();
-    const useEffect = jest.fn();
-    const useState = jest.fn<UseStateResult, UseStateArgs>()
-      .mockReturnValue([2, setStateId]);
+    mockUseRef({
+      isInit,
+      optionsCache,
+    });
 
-    useAsyncPaginateBasePure(
-      makeUseRef({
-        isInit,
-        optionsCache,
-      }),
-      (useState as unknown as typeof reactUseState),
-      useEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      defaultRequestOptions,
+    useAsyncPaginateBase(
       defaultParams,
     );
 
-    useEffect.mock.calls[0][0]();
+    (useEffect as jest.Mock).mock.calls[0][0]();
 
     expect(isInit.current).toBe(false);
     expect(optionsCache.current).toEqual({});
     expect(setStateId).toHaveBeenCalledTimes(1);
-    expect(setStateId.mock.calls[0][0]).toEqual(increaseStateId);
+    expect(setStateId).toHaveBeenCalledWith(increaseStateId);
   });
 
   test('should provide inputValue as dependency to second useEffect', async () => {
-    const useEffect = jest.fn();
-
-    useAsyncPaginateBasePure(
-      makeUseRef({}),
-      (defaultUseState as unknown as typeof reactUseState),
-      useEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      defaultRequestOptions,
+    useAsyncPaginateBase(
       {
         ...defaultParams,
         inputValue: 'test',
       },
     );
 
-    expect(useEffect.mock.calls[1][1]).toEqual(['test']);
+    expect((useEffect as jest.Mock).mock.calls[1][1]).toEqual(['test']);
   });
 
   test('should load options on inputValue change if options are not cached if menu is open', async () => {
-    const requestOptionsParam = jest.fn();
-    const useEffect = jest.fn();
-
-    useAsyncPaginateBasePure(
-      makeUseRef({
-        optionsCache: {
-          current: {},
-        },
-      }),
-      (defaultUseState as unknown as typeof reactUseState),
-      useEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      requestOptionsParam,
+    useAsyncPaginateBase(
       {
         ...defaultParams,
         inputValue: 'test',
@@ -549,27 +304,13 @@ describe('useAsyncPaginateBasePure', () => {
       },
     );
 
-    useEffect.mock.calls[1][0]();
+    (useEffect as jest.Mock).mock.calls[1][0]();
 
-    expect(requestOptionsParam).toHaveBeenCalledTimes(1);
+    expect(requestOptions).toHaveBeenCalledTimes(1);
   });
 
   test('should not load options on inputValue change if options are not cached if menu is not open', async () => {
-    const requestOptionsParam = jest.fn();
-    const useEffect = jest.fn();
-
-    useAsyncPaginateBasePure(
-      makeUseRef({
-        optionsCache: {
-          current: {},
-        },
-      }),
-      (defaultUseState as unknown as typeof reactUseState),
-      useEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      requestOptionsParam,
+    useAsyncPaginateBase(
       {
         ...defaultParams,
         inputValue: 'test',
@@ -577,152 +318,99 @@ describe('useAsyncPaginateBasePure', () => {
       },
     );
 
-    useEffect.mock.calls[1][0]();
+    (useEffect as jest.Mock).mock.calls[1][0]();
 
-    expect(requestOptionsParam).toHaveBeenCalledTimes(0);
+    expect(requestOptions).toHaveBeenCalledTimes(0);
   });
 
   test('should not load options on inputValue change if options are cached', async () => {
-    const requestOptionsParam = jest.fn();
-    const useEffect = jest.fn();
+    mockUseRef({
+      optionsCache: {
+        current: {
+          test: {
+            options: [
+              {
+                value: 1,
+                label: '1',
+              },
 
-    useAsyncPaginateBasePure(
-      makeUseRef({
-        optionsCache: {
-          current: {
-            test: {
-              options: [
-                {
-                  value: 1,
-                  label: '1',
-                },
-
-                {
-                  value: 2,
-                  label: '2',
-                },
-              ],
-              hasMore: true,
-              isLoading: false,
-              isFirstLoad: false,
-            },
+              {
+                value: 2,
+                label: '2',
+              },
+            ],
+            hasMore: true,
+            isLoading: false,
+            isFirstLoad: false,
           },
         },
-      }),
-      (defaultUseState as unknown as typeof reactUseState),
-      useEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      requestOptionsParam,
+      },
+    });
+
+    useAsyncPaginateBase(
       {
         ...defaultParams,
         inputValue: 'test',
       },
     );
 
-    useEffect.mock.calls[1][0]();
+    (useEffect as jest.Mock).mock.calls[1][0]();
 
-    expect(requestOptionsParam).toHaveBeenCalledTimes(0);
+    expect(requestOptions).toHaveBeenCalledTimes(0);
   });
 
   test('should provide menuIsOpen as dependency to third useEffect', async () => {
-    const useEffect = jest.fn();
-
-    useAsyncPaginateBasePure(
-      makeUseRef({}),
-      (defaultUseState as unknown as typeof reactUseState),
-      useEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      defaultRequestOptions,
+    useAsyncPaginateBase(
       {
         ...defaultParams,
         menuIsOpen: true,
       },
     );
 
-    expect(useEffect.mock.calls[2][1]).toEqual([true]);
+    expect((useEffect as jest.Mock).mock.calls[2][1]).toEqual([true]);
   });
 
   test('should not load options from third useEffect if menuIsOpen is false', async () => {
-    const requestOptionsParam = jest.fn();
-    const useEffect = jest.fn();
-
-    useAsyncPaginateBasePure(
-      makeUseRef({
-        optionsCache: {
-          current: {},
-        },
-      }),
-      (defaultUseState as unknown as typeof reactUseState),
-      useEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      requestOptionsParam,
+    useAsyncPaginateBase(
       {
         ...defaultParams,
         menuIsOpen: false,
       },
     );
 
-    useEffect.mock.calls[2][0]();
+    (useEffect as jest.Mock).mock.calls[2][0]();
 
-    expect(requestOptionsParam).toHaveBeenCalledTimes(0);
+    expect(requestOptions).toHaveBeenCalledTimes(0);
   });
 
   test('should not load options from third useEffect if optionsCache defined for empty search', async () => {
-    const requestOptionsParam = jest.fn();
-    const useEffect = jest.fn();
-
-    useAsyncPaginateBasePure(
-      makeUseRef({
-        optionsCache: {
-          current: {
-            '': {
-              options: [],
-              hasMore: true,
-              isLoading: false,
-              isFirstLoad: false,
-            },
+    mockUseRef({
+      optionsCache: {
+        current: {
+          '': {
+            options: [],
+            hasMore: true,
+            isLoading: false,
+            isFirstLoad: false,
           },
         },
-      }),
-      (defaultUseState as unknown as typeof reactUseState),
-      useEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      requestOptionsParam,
+      },
+    });
+
+    useAsyncPaginateBase(
       {
         ...defaultParams,
         menuIsOpen: true,
       },
     );
 
-    useEffect.mock.calls[2][0]();
+    (useEffect as jest.Mock).mock.calls[2][0]();
 
-    expect(requestOptionsParam).toHaveBeenCalledTimes(0);
+    expect(requestOptions).toHaveBeenCalledTimes(0);
   });
 
   test('should not load options from third useEffect if loadOptionsOnMenuOpen is false', async () => {
-    const requestOptionsParam = jest.fn();
-    const useEffect = jest.fn();
-
-    useAsyncPaginateBasePure(
-      makeUseRef({
-        optionsCache: {
-          current: {},
-        },
-      }),
-      (defaultUseState as unknown as typeof reactUseState),
-      useEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      requestOptionsParam,
+    useAsyncPaginateBase(
       {
         ...defaultParams,
         loadOptionsOnMenuOpen: false,
@@ -730,53 +418,26 @@ describe('useAsyncPaginateBasePure', () => {
       },
     );
 
-    useEffect.mock.calls[2][0]();
+    (useEffect as jest.Mock).mock.calls[2][0]();
 
-    expect(requestOptionsParam).toHaveBeenCalledTimes(0);
+    expect(requestOptions).toHaveBeenCalledTimes(0);
   });
 
   test('should load options from third useEffect', async () => {
-    const requestOptionsParam = jest.fn();
-    const useEffect = jest.fn();
-
-    useAsyncPaginateBasePure(
-      makeUseRef({
-        optionsCache: {
-          current: {},
-        },
-      }),
-      (defaultUseState as unknown as typeof reactUseState),
-      useEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      requestOptionsParam,
+    useAsyncPaginateBase(
       {
         ...defaultParams,
         menuIsOpen: true,
       },
     );
 
-    useEffect.mock.calls[2][0]();
+    (useEffect as jest.Mock).mock.calls[2][0]();
 
-    expect(requestOptionsParam).toHaveBeenCalledTimes(1);
+    expect(requestOptions).toHaveBeenCalledTimes(1);
   });
 
   test('should not load options on scroll to bottom if cache not defined for current search', async () => {
-    const requestOptionsParam = jest.fn();
-
-    const result = useAsyncPaginateBasePure(
-      makeUseRef({
-        optionsCache: {
-          current: {},
-        },
-      }),
-      (defaultUseState as unknown as typeof reactUseState),
-      defaultUseEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      requestOptionsParam,
+    const result = useAsyncPaginateBase(
       {
         ...defaultParams,
         inputValue: 'test',
@@ -785,31 +446,24 @@ describe('useAsyncPaginateBasePure', () => {
 
     result.handleScrolledToBottom();
 
-    expect(requestOptionsParam).toHaveBeenCalledTimes(0);
+    expect(requestOptions).toHaveBeenCalledTimes(0);
   });
 
   test('should load options on scroll to bottom if cache defined for current search', async () => {
-    const requestOptionsParam = jest.fn();
-
-    const result = useAsyncPaginateBasePure(
-      makeUseRef({
-        optionsCache: {
-          current: {
-            test: {
-              options: [],
-              hasMore: true,
-              isLoading: false,
-              isFirstLoad: false,
-            },
+    mockUseRef({
+      optionsCache: {
+        current: {
+          test: {
+            options: [],
+            hasMore: true,
+            isLoading: false,
+            isFirstLoad: false,
           },
         },
-      }),
-      (defaultUseState as unknown as typeof reactUseState),
-      defaultUseEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      requestOptionsParam,
+      },
+    });
+
+    const result = useAsyncPaginateBase(
       {
         ...defaultParams,
         inputValue: 'test',
@@ -818,18 +472,11 @@ describe('useAsyncPaginateBasePure', () => {
 
     result.handleScrolledToBottom();
 
-    expect(requestOptionsParam).toHaveBeenCalledTimes(1);
+    expect(requestOptions).toHaveBeenCalledTimes(1);
   });
 
   test('should provide default shouldLoadMore', async () => {
-    const result = useAsyncPaginateBasePure(
-      makeUseRef({}),
-      (defaultUseState as unknown as typeof reactUseState),
-      defaultUseEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      defaultRequestOptions,
+    const result = useAsyncPaginateBase(
       defaultParams,
     );
 
@@ -839,14 +486,7 @@ describe('useAsyncPaginateBasePure', () => {
   test('should provide redefined shouldLoadMore', async () => {
     const shouldLoadMore = jest.fn();
 
-    const result = useAsyncPaginateBasePure(
-      makeUseRef({}),
-      (defaultUseState as unknown as typeof reactUseState),
-      defaultUseEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      defaultRequestOptions,
+    const result = useAsyncPaginateBase(
       {
         ...defaultParams,
         shouldLoadMore,
@@ -857,14 +497,7 @@ describe('useAsyncPaginateBasePure', () => {
   });
 
   test('should provide default filterOption', async () => {
-    const result = useAsyncPaginateBasePure(
-      makeUseRef({}),
-      (defaultUseState as unknown as typeof reactUseState),
-      defaultUseEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      defaultRequestOptions,
+    const result = useAsyncPaginateBase(
       defaultParams,
     );
 
@@ -874,14 +507,7 @@ describe('useAsyncPaginateBasePure', () => {
   test('should provide redefined filterOption', async () => {
     const filterOption = jest.fn();
 
-    const result = useAsyncPaginateBasePure(
-      makeUseRef({}),
-      (defaultUseState as unknown as typeof reactUseState),
-      defaultUseEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      defaultRequestOptions,
+    const result = useAsyncPaginateBase(
       {
         ...defaultParams,
         filterOption,
@@ -892,18 +518,7 @@ describe('useAsyncPaginateBasePure', () => {
   });
 
   test('should provide initial params if cache not defined for current search', async () => {
-    const result = useAsyncPaginateBasePure(
-      makeUseRef({
-        optionsCache: {
-          current: {},
-        },
-      }),
-      (defaultUseState as unknown as typeof reactUseState),
-      defaultUseEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      defaultRequestOptions,
+    const result = useAsyncPaginateBase(
       {
         ...defaultParams,
         inputValue: 'test',
@@ -928,25 +543,20 @@ describe('useAsyncPaginateBasePure', () => {
       },
     ];
 
-    const result = useAsyncPaginateBasePure(
-      makeUseRef({
-        optionsCache: {
-          current: {
-            test: {
-              options,
-              hasMore: true,
-              isLoading: true,
-              isFirstLoad: false,
-            },
+    mockUseRef({
+      optionsCache: {
+        current: {
+          test: {
+            options,
+            hasMore: true,
+            isLoading: true,
+            isFirstLoad: false,
           },
         },
-      }),
-      (defaultUseState as unknown as typeof reactUseState),
-      defaultUseEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      defaultRequestOptions,
+      },
+    });
+
+    const result = useAsyncPaginateBase(
       {
         ...defaultParams,
         inputValue: 'test',
@@ -959,27 +569,20 @@ describe('useAsyncPaginateBasePure', () => {
   });
 
   test('should provide default reduceOptions to requestOptions', async () => {
-    const requestOptionsParam = jest.fn();
-
-    const result = useAsyncPaginateBasePure(
-      makeUseRef({
-        optionsCache: {
-          current: {
-            test: {
-              options: [],
-              hasMore: true,
-              isLoading: false,
-              isFirstLoad: false,
-            },
+    mockUseRef({
+      optionsCache: {
+        current: {
+          test: {
+            options: [],
+            hasMore: true,
+            isLoading: false,
+            isFirstLoad: false,
           },
         },
-      }),
-      (defaultUseState as unknown as typeof reactUseState),
-      defaultUseEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      requestOptionsParam,
+      },
+    });
+
+    const result = useAsyncPaginateBase(
       {
         ...defaultParams,
         inputValue: 'test',
@@ -988,33 +591,27 @@ describe('useAsyncPaginateBasePure', () => {
 
     result.handleScrolledToBottom();
 
-    expect(requestOptionsParam).toHaveBeenCalledTimes(1);
-    expect(requestOptionsParam.mock.calls[0][7]).toBe(defaultReduceOptions);
+    expect(requestOptions).toHaveBeenCalledTimes(1);
+    expect((requestOptions as jest.Mock).mock.calls[0][5]).toBe(defaultReduceOptions);
   });
 
   test('should provide redefined reduceOptions to requestOptions', async () => {
-    const requestOptionsParam = jest.fn();
     const reduceOptions = jest.fn();
 
-    const result = useAsyncPaginateBasePure(
-      makeUseRef({
-        optionsCache: {
-          current: {
-            test: {
-              options: [],
-              hasMore: true,
-              isLoading: false,
-              isFirstLoad: false,
-            },
+    mockUseRef({
+      optionsCache: {
+        current: {
+          test: {
+            options: [],
+            hasMore: true,
+            isLoading: false,
+            isFirstLoad: false,
           },
         },
-      }),
-      (defaultUseState as unknown as typeof reactUseState),
-      defaultUseEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      requestOptionsParam,
+      },
+    });
+
+    const result = useAsyncPaginateBase(
       {
         ...defaultParams,
         inputValue: 'test',
@@ -1024,21 +621,18 @@ describe('useAsyncPaginateBasePure', () => {
 
     result.handleScrolledToBottom();
 
-    expect(requestOptionsParam).toHaveBeenCalledTimes(1);
-    expect(requestOptionsParam.mock.calls[0][7]).toBe(reduceOptions);
+    expect(requestOptions).toHaveBeenCalledTimes(1);
+    expect((requestOptions as jest.Mock).mock.calls[0][5]).toBe(reduceOptions);
   });
 
   test('should reduce change cached options and set next increase state id if mounted', async () => {
-    const requestOptionsParam = jest.fn();
-
-    const setStateId = jest.fn();
-    const useState = jest.fn<UseStateResult, UseStateArgs>()
-      .mockReturnValue([2, setStateId]);
-
     const reduceState = jest.fn()
       .mockReturnValue({
         test2: defaultCacheItem,
       });
+
+    const setStateId = jest.fn();
+    (useState as jest.Mock).mockReturnValue([1, setStateId]);
 
     const optionsCache = {
       current: {
@@ -1046,20 +640,15 @@ describe('useAsyncPaginateBasePure', () => {
       },
     };
 
+    mockUseRef({
+      optionsCache,
+    });
+
     (useIsMountedRef as jest.Mock).mockReturnValue({
       current: true,
     });
 
-    const result = useAsyncPaginateBasePure(
-      makeUseRef({
-        optionsCache,
-      }),
-      (useState as unknown as typeof reactUseState),
-      defaultUseEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      requestOptionsParam,
+    const result = useAsyncPaginateBase(
       {
         ...defaultParams,
         inputValue: 'test1',
@@ -1068,7 +657,7 @@ describe('useAsyncPaginateBasePure', () => {
 
     result.handleScrolledToBottom();
 
-    requestOptionsParam.mock.calls[0][5](reduceState);
+    (requestOptions as jest.Mock).mock.calls[0][4](reduceState);
 
     expect(reduceState).toHaveBeenCalledTimes(1);
     expect(reduceState).toHaveBeenCalledWith({
@@ -1084,16 +673,13 @@ describe('useAsyncPaginateBasePure', () => {
   });
 
   test('should reduce change cached options and not set next increase state id if not mounted', async () => {
-    const requestOptionsParam = jest.fn();
-
-    const setStateId = jest.fn();
-    const useState = jest.fn<UseStateResult, UseStateArgs>()
-      .mockReturnValue([2, setStateId]);
-
     const reduceState = jest.fn()
       .mockReturnValue({
         test2: defaultCacheItem,
       });
+
+    const setStateId = jest.fn();
+    (useState as jest.Mock).mockReturnValue([1, setStateId]);
 
     const optionsCache = {
       current: {
@@ -1101,16 +687,11 @@ describe('useAsyncPaginateBasePure', () => {
       },
     };
 
-    const result = useAsyncPaginateBasePure(
-      makeUseRef({
-        optionsCache,
-      }),
-      (useState as unknown as typeof reactUseState),
-      defaultUseEffect,
-      defaultUseCallback,
-      defaultValidateResponse,
-      getInitialOptionsCache,
-      requestOptionsParam,
+    mockUseRef({
+      optionsCache,
+    });
+
+    const result = useAsyncPaginateBase(
       {
         ...defaultParams,
         inputValue: 'test1',
@@ -1119,7 +700,7 @@ describe('useAsyncPaginateBasePure', () => {
 
     result.handleScrolledToBottom();
 
-    requestOptionsParam.mock.calls[0][5](reduceState);
+    (requestOptions as jest.Mock).mock.calls[0][4](reduceState);
 
     expect(reduceState).toHaveBeenCalledTimes(1);
     expect(reduceState).toHaveBeenCalledWith({
@@ -1131,940 +712,5 @@ describe('useAsyncPaginateBasePure', () => {
     });
 
     expect(setStateId).toHaveBeenCalledTimes(0);
-  });
-});
-
-describe('requestOptions', () => {
-  const defaultParamsRef = {
-    current: defaultParams,
-  };
-
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore
-  const defaultSleep: typeof sleepLib = async (): Promise<void> => {};
-
-  const defaultSetOptionsCache = (): void => {};
-
-  test('should request if options not cached', async () => {
-    const newOptions = [
-      {
-        value: 1,
-        label: '1',
-      },
-
-      {
-        value: 2,
-        label: '2',
-      },
-    ];
-
-    const setOptionsCache = jest.fn();
-    const loadOptions = jest.fn()
-      .mockReturnValue({
-        options: newOptions,
-        hasMore: true,
-      });
-
-    const additional = Symbol('additional');
-
-    await requestOptions(
-      'autoload',
-      {
-        ...defaultParamsRef,
-        current: {
-          ...defaultParams,
-          loadOptions,
-          inputValue: 'test',
-          additional,
-        },
-      },
-      {
-        current: {},
-      },
-      0,
-      defaultSleep,
-      setOptionsCache,
-      defaultValidateResponse,
-      defaultReduceOptions,
-    );
-
-    expect(loadOptions).toHaveBeenCalledTimes(1);
-    expect(loadOptions.mock.calls[0][0]).toBe('test');
-    expect(loadOptions.mock.calls[0][1]).toEqual([]);
-    expect(loadOptions.mock.calls[0][2]).toEqual(additional);
-
-    expect(setOptionsCache).toHaveBeenCalledTimes(2);
-
-    const intermediateCache = setOptionsCache.mock.calls[0][0]({});
-
-    expect(intermediateCache).toEqual({
-      test: {
-        isFirstLoad: true,
-        options: [],
-        hasMore: true,
-        isLoading: true,
-        additional,
-      },
-    });
-
-    const lastCache = setOptionsCache.mock.calls[1][0](intermediateCache);
-
-    expect(lastCache).toEqual({
-      test: {
-        isFirstLoad: false,
-        options: newOptions,
-        hasMore: true,
-        isLoading: false,
-        additional,
-      },
-    });
-  });
-
-  test('should request if options cached', async () => {
-    const prevOptions = [
-      {
-        value: 1,
-        label: '1',
-      },
-
-      {
-        value: 2,
-        label: '2',
-      },
-    ];
-
-    const newOptions = [
-      {
-        value: 3,
-        label: '4',
-      },
-
-      {
-        value: 4,
-        label: '4',
-      },
-    ];
-
-    const setOptionsCache = jest.fn();
-    const loadOptions = jest.fn()
-      .mockReturnValue({
-        options: newOptions,
-        hasMore: true,
-      });
-
-    const additional = Symbol('additional');
-
-    const initialOptionsCache = {
-      test: {
-        options: prevOptions,
-        hasMore: true,
-        isLoading: false,
-        isFirstLoad: false,
-        additional,
-      },
-    };
-
-    await requestOptions(
-      'autoload',
-      {
-        ...defaultParamsRef,
-        current: {
-          ...defaultParams,
-          loadOptions,
-          inputValue: 'test',
-        },
-      },
-      {
-        current: initialOptionsCache,
-      },
-      0,
-      defaultSleep,
-      setOptionsCache,
-      defaultValidateResponse,
-      defaultReduceOptions,
-    );
-
-    expect(loadOptions).toHaveBeenCalledTimes(1);
-    expect(loadOptions.mock.calls[0][0]).toBe('test');
-    expect(loadOptions.mock.calls[0][1]).toBe(prevOptions);
-    expect(loadOptions.mock.calls[0][2]).toEqual(additional);
-
-    expect(setOptionsCache).toHaveBeenCalledTimes(2);
-
-    const intermediateCache = setOptionsCache.mock.calls[0][0](initialOptionsCache);
-
-    expect(intermediateCache).toEqual({
-      test: {
-        isFirstLoad: false,
-        options: prevOptions,
-        hasMore: true,
-        isLoading: true,
-        additional,
-      },
-    });
-
-    const lastCache = setOptionsCache.mock.calls[1][0](intermediateCache);
-
-    expect(lastCache).toEqual({
-      test: {
-        isFirstLoad: false,
-        options: [
-          ...prevOptions,
-          ...newOptions,
-        ],
-        hasMore: true,
-        isLoading: false,
-        additional,
-      },
-    });
-  });
-
-  test('should not request if options are loading for current search', async () => {
-    const setOptionsCache = jest.fn();
-    const loadOptions = jest.fn()
-      .mockReturnValue({
-        options: [],
-        hasMore: true,
-      });
-
-    const additional = Symbol('additional');
-
-    await requestOptions(
-      'autoload',
-      {
-        ...defaultParamsRef,
-        current: {
-          ...defaultParams,
-          loadOptions,
-          inputValue: 'test',
-          additional,
-        },
-      },
-      {
-        current: {
-          test: {
-            options: [],
-            hasMore: true,
-            isLoading: true,
-            isFirstLoad: false,
-          },
-        },
-      },
-      0,
-      defaultSleep,
-      setOptionsCache,
-      defaultValidateResponse,
-      defaultReduceOptions,
-    );
-
-    expect(loadOptions).toHaveBeenCalledTimes(0);
-    expect(setOptionsCache).toHaveBeenCalledTimes(0);
-  });
-
-  test('should not request if hasMore is false for current search', async () => {
-    const setOptionsCache = jest.fn();
-    const loadOptions = jest.fn()
-      .mockReturnValue({
-        options: [],
-        hasMore: true,
-      });
-
-    const additional = Symbol('additional');
-
-    await requestOptions(
-      'autoload',
-      {
-        ...defaultParamsRef,
-        current: {
-          ...defaultParams,
-          loadOptions,
-          inputValue: 'test',
-          additional,
-        },
-      },
-      {
-        current: {
-          test: {
-            options: [],
-            hasMore: false,
-            isLoading: false,
-            isFirstLoad: false,
-          },
-        },
-      },
-      0,
-      defaultSleep,
-      setOptionsCache,
-      defaultValidateResponse,
-      defaultReduceOptions,
-    );
-
-    expect(loadOptions).toHaveBeenCalledTimes(0);
-    expect(setOptionsCache).toHaveBeenCalledTimes(0);
-  });
-
-  test('should request with error', async () => {
-    const setOptionsCache = jest.fn();
-    const loadOptions = jest.fn()
-      .mockRejectedValue(new Error());
-
-    const additional = Symbol('additional');
-
-    await requestOptions(
-      'autoload',
-      {
-        ...defaultParamsRef,
-        current: {
-          ...defaultParams,
-          loadOptions,
-          inputValue: 'test',
-          additional,
-        },
-      },
-      {
-        current: {},
-      },
-      0,
-      defaultSleep,
-      setOptionsCache,
-      defaultValidateResponse,
-      defaultReduceOptions,
-    );
-
-    expect(loadOptions).toHaveBeenCalledTimes(1);
-    expect(loadOptions).toHaveBeenCalledWith('test', [], additional);
-
-    expect(setOptionsCache).toHaveBeenCalledTimes(2);
-
-    const intermediateCache = setOptionsCache.mock.calls[0][0]({});
-
-    expect(intermediateCache).toEqual({
-      test: {
-        isFirstLoad: true,
-        options: [],
-        hasMore: true,
-        isLoading: true,
-        additional,
-      },
-    });
-
-    const lastCache = setOptionsCache.mock.calls[1][0](intermediateCache);
-
-    expect(lastCache).toEqual({
-      test: {
-        isFirstLoad: true,
-        options: [],
-        hasMore: true,
-        isLoading: false,
-        additional,
-      },
-    });
-  });
-
-  test('should redefine reduceOptions', async () => {
-    const prevOptions = [
-      {
-        value: 1,
-        label: '1',
-      },
-
-      {
-        value: 2,
-        label: '2',
-      },
-    ];
-
-    const newOptions = [
-      {
-        value: 3,
-        label: '4',
-      },
-
-      {
-        value: 4,
-        label: '4',
-      },
-    ];
-
-    const reducedOptions = [
-      {
-        value: 5,
-        label: '5',
-      },
-    ];
-
-    const reduceOptions = jest.fn()
-      .mockReturnValue(reducedOptions);
-
-    const setOptionsCache = jest.fn();
-    const loadOptions = jest.fn()
-      .mockReturnValue({
-        options: newOptions,
-        hasMore: true,
-      });
-
-    const additional = Symbol('additional');
-
-    const initialOptionsCache = {
-      test: {
-        options: prevOptions,
-        hasMore: true,
-        isLoading: false,
-        isFirstLoad: false,
-        additional,
-      },
-    };
-
-    await requestOptions(
-      'autoload',
-      {
-        ...defaultParamsRef,
-        current: {
-          ...defaultParams,
-          loadOptions,
-          inputValue: 'test',
-        },
-      },
-      {
-        current: initialOptionsCache,
-      },
-      0,
-      defaultSleep,
-      setOptionsCache,
-      defaultValidateResponse,
-      reduceOptions,
-    );
-
-    expect(loadOptions).toHaveBeenCalledTimes(1);
-    expect(loadOptions.mock.calls[0][0]).toBe('test');
-    expect(loadOptions.mock.calls[0][1]).toBe(prevOptions);
-    expect(loadOptions.mock.calls[0][2]).toEqual(additional);
-
-    expect(setOptionsCache).toHaveBeenCalledTimes(2);
-
-    const intermediateCache = setOptionsCache.mock.calls[0][0](initialOptionsCache);
-
-    expect(intermediateCache).toEqual({
-      test: {
-        isFirstLoad: false,
-        options: prevOptions,
-        hasMore: true,
-        isLoading: true,
-        additional,
-      },
-    });
-
-    const lastCache = setOptionsCache.mock.calls[1][0](intermediateCache);
-
-    expect(lastCache).toEqual({
-      test: {
-        isFirstLoad: false,
-        options: reducedOptions,
-        hasMore: true,
-        isLoading: false,
-        additional,
-      },
-    });
-
-    expect(reduceOptions).toHaveBeenCalledTimes(1);
-    expect(reduceOptions.mock.calls[0][0]).toBe(prevOptions);
-    expect(reduceOptions.mock.calls[0][1]).toBe(newOptions);
-    expect(reduceOptions.mock.calls[0][2]).toBe(additional);
-  });
-
-  test('should validate response', async () => {
-    const newOptions = [
-      {
-        value: 1,
-        label: '1',
-      },
-
-      {
-        value: 2,
-        label: '2',
-      },
-    ];
-
-    const response = {
-      options: newOptions,
-      hasMore: true,
-    };
-
-    const validateResponseParam = jest.fn<void, [Console, any]>(() => {
-      throw new Error();
-    });
-
-    const loadOptions = jest.fn()
-      .mockReturnValue(response);
-
-    let hasError = false;
-    try {
-      await requestOptions(
-        'autoload',
-        {
-          ...defaultParamsRef,
-          current: {
-            ...defaultParams,
-            loadOptions,
-          },
-        },
-        {
-          current: {},
-        },
-        0,
-        defaultSleep,
-        defaultSetOptionsCache,
-        validateResponseParam,
-        defaultReduceOptions,
-      );
-    } catch (e) {
-      hasError = true;
-    }
-
-    expect(hasError).toBe(true);
-
-    expect(validateResponseParam).toHaveBeenCalledTimes(1);
-    expect(validateResponseParam.mock.calls[0][1]).toBe(response);
-  });
-
-  test('should not sleep if debounceTimeout is 0', async () => {
-    const sleep = jest.fn();
-
-    await requestOptions(
-      'input-change',
-      {
-        ...defaultParamsRef,
-        current: {
-          ...defaultParams,
-        },
-      },
-      {
-        current: {},
-      },
-      0,
-      sleep,
-      defaultSetOptionsCache,
-      defaultValidateResponse,
-      defaultReduceOptions,
-    );
-
-    expect(sleep).toHaveBeenCalledTimes(0);
-  });
-
-  test('should not sleep if debounceTimeout bigger than 0 and caller is not "input-change"', async () => {
-    const sleep = jest.fn();
-
-    await requestOptions(
-      'autoload',
-      {
-        ...defaultParamsRef,
-        current: {
-          ...defaultParams,
-        },
-      },
-      {
-        current: {},
-      },
-      0,
-      sleep,
-      defaultSetOptionsCache,
-      defaultValidateResponse,
-      defaultReduceOptions,
-    );
-
-    expect(sleep).toHaveBeenCalledTimes(0);
-  });
-
-  test('should sleep if debounceTimeout bigger than 0 and caller is "input-change"', async () => {
-    const sleep = jest.fn();
-
-    const newOptions = [
-      {
-        value: 1,
-        label: '1',
-      },
-
-      {
-        value: 2,
-        label: '2',
-      },
-    ];
-
-    const setOptionsCache = jest.fn();
-    const loadOptions = jest.fn()
-      .mockReturnValue({
-        options: newOptions,
-        hasMore: true,
-      });
-
-    const additional = Symbol('additional');
-
-    await requestOptions(
-      'input-change',
-      {
-        ...defaultParamsRef,
-        current: {
-          ...defaultParams,
-          loadOptions,
-          inputValue: 'test',
-          additional,
-        },
-      },
-      {
-        current: {},
-      },
-      1234,
-      sleep,
-      setOptionsCache,
-      defaultValidateResponse,
-      defaultReduceOptions,
-    );
-
-    expect(sleep).toHaveBeenCalledTimes(1);
-    expect(sleep.mock.calls[0][0]).toBe(1234);
-
-    expect(loadOptions).toHaveBeenCalledTimes(1);
-    expect(loadOptions).toHaveBeenCalledWith('test', [], additional);
-
-    expect(setOptionsCache).toHaveBeenCalledTimes(2);
-
-    const intermediateCache = setOptionsCache.mock.calls[0][0]({});
-
-    expect(intermediateCache).toEqual({
-      test: {
-        isFirstLoad: true,
-        options: [],
-        hasMore: true,
-        isLoading: true,
-        additional,
-      },
-    });
-
-    const lastCache = setOptionsCache.mock.calls[1][0](intermediateCache);
-
-    expect(lastCache).toEqual({
-      test: {
-        isFirstLoad: false,
-        options: newOptions,
-        hasMore: true,
-        isLoading: false,
-        additional,
-      },
-    });
-  });
-
-  test('should cancel loading if inputValue has changed during sleep for empty cache', async () => {
-    const newOptions = [
-      {
-        value: 1,
-        label: '1',
-      },
-
-      {
-        value: 2,
-        label: '2',
-      },
-    ];
-
-    const setOptionsCache = jest.fn();
-    const loadOptions = jest.fn()
-      .mockReturnValue({
-        options: newOptions,
-        hasMore: true,
-      });
-
-    const additional = Symbol('additional');
-
-    const paramsRef = {
-      ...defaultParamsRef,
-      current: {
-        ...defaultParams,
-        loadOptions,
-        inputValue: 'test',
-        additional,
-      },
-    };
-
-    await requestOptions(
-      'input-change',
-      paramsRef,
-      {
-        current: {},
-      },
-      1234,
-      (async () => {
-        paramsRef.current.inputValue = 'test2';
-      }) as unknown as typeof sleepLib,
-      setOptionsCache,
-      defaultValidateResponse,
-      defaultReduceOptions,
-    );
-
-    expect(loadOptions).toHaveBeenCalledTimes(0);
-
-    expect(setOptionsCache).toHaveBeenCalledTimes(2);
-
-    const intermediateCache = setOptionsCache.mock.calls[0][0]({});
-
-    expect(intermediateCache).toEqual({
-      test: {
-        isFirstLoad: true,
-        options: [],
-        hasMore: true,
-        isLoading: true,
-        additional,
-      },
-    });
-
-    const lastCache = setOptionsCache.mock.calls[1][0](intermediateCache);
-
-    expect(lastCache).toEqual({});
-  });
-
-  test('should cancel loading if inputValue has changed during sleep for filled cache', async () => {
-    const newOptions = [
-      {
-        value: 1,
-        label: '1',
-      },
-
-      {
-        value: 2,
-        label: '2',
-      },
-    ];
-
-    const setOptionsCache = jest.fn();
-    const loadOptions = jest.fn()
-      .mockReturnValue({
-        options: newOptions,
-        hasMore: true,
-      });
-
-    const additional = Symbol('additional');
-
-    const paramsRef = {
-      ...defaultParamsRef,
-      current: {
-        ...defaultParams,
-        loadOptions,
-        inputValue: 'test',
-        additional,
-      },
-    };
-
-    await requestOptions(
-      'input-change',
-      paramsRef,
-      {
-        current: {
-          test: {
-            isFirstLoad: true,
-            options: [],
-            hasMore: true,
-            isLoading: false,
-            additional,
-          },
-        },
-      },
-      1234,
-      (async () => {
-        paramsRef.current.inputValue = 'test2';
-      }) as unknown as typeof sleepLib,
-      setOptionsCache,
-      defaultValidateResponse,
-      defaultReduceOptions,
-    );
-
-    expect(loadOptions).toHaveBeenCalledTimes(0);
-
-    expect(setOptionsCache).toHaveBeenCalledTimes(2);
-
-    const intermediateCache = setOptionsCache.mock.calls[0][0]({});
-
-    expect(intermediateCache).toEqual({
-      test: {
-        isFirstLoad: true,
-        options: [],
-        hasMore: true,
-        isLoading: true,
-        additional,
-      },
-    });
-
-    const lastCache = setOptionsCache.mock.calls[1][0](intermediateCache);
-
-    expect(lastCache).toEqual({
-      test: {
-        isFirstLoad: true,
-        options: [],
-        hasMore: true,
-        isLoading: false,
-        additional,
-      },
-    });
-  });
-
-  test('should redefine additional with response', async () => {
-    const additional1 = Symbol('additional1');
-    const additional2 = Symbol('additional2');
-
-    const setOptionsCache = jest.fn();
-    const loadOptions = jest.fn()
-      .mockReturnValue({
-        options: [],
-        hasMore: true,
-        additional: additional2,
-      });
-
-    await requestOptions(
-      'autoload',
-      {
-        ...defaultParamsRef,
-        current: {
-          ...defaultParams,
-          loadOptions,
-          inputValue: 'test',
-          additional: additional1,
-        },
-      },
-      {
-        current: {},
-      },
-      0,
-      defaultSleep,
-      setOptionsCache,
-      defaultValidateResponse,
-      defaultReduceOptions,
-    );
-
-    expect(setOptionsCache).toHaveBeenCalledTimes(2);
-
-    const lastCache = setOptionsCache.mock.calls[1][0](
-      setOptionsCache.mock.calls[0][0](
-        {},
-      ),
-    );
-
-    expect(lastCache.test.additional).toBe(additional2);
-  });
-
-  test('should not redefine additional with response', async () => {
-    const additional1 = Symbol('additional1');
-
-    const setOptionsCache = jest.fn();
-    const loadOptions = jest.fn()
-      .mockReturnValue({
-        options: [],
-        hasMore: true,
-      });
-
-    await requestOptions(
-      'autoload',
-      {
-        ...defaultParamsRef,
-        current: {
-          ...defaultParams,
-          loadOptions,
-          inputValue: 'test',
-          additional: additional1,
-        },
-      },
-      {
-        current: {},
-      },
-      0,
-      defaultSleep,
-      setOptionsCache,
-      defaultValidateResponse,
-      defaultReduceOptions,
-    );
-
-    expect(setOptionsCache).toHaveBeenCalledTimes(2);
-
-    const lastCache = setOptionsCache.mock.calls[1][0](
-      setOptionsCache.mock.calls[0][0](
-        {},
-      ),
-    );
-
-    expect(lastCache.test.additional).toBe(additional1);
-  });
-
-  test('should set truthy hasMore with response', async () => {
-    const setOptionsCache = jest.fn();
-    const loadOptions = jest.fn()
-      .mockReturnValue({
-        options: [],
-        hasMore: true,
-      });
-
-    await requestOptions(
-      'autoload',
-      {
-        ...defaultParamsRef,
-        current: {
-          ...defaultParams,
-          loadOptions,
-          inputValue: 'test',
-        },
-      },
-      {
-        current: {},
-      },
-      0,
-      defaultSleep,
-      setOptionsCache,
-      defaultValidateResponse,
-      defaultReduceOptions,
-    );
-
-    expect(setOptionsCache).toHaveBeenCalledTimes(2);
-
-    const lastCache = setOptionsCache.mock.calls[1][0](
-      setOptionsCache.mock.calls[0][0](
-        {},
-      ),
-    );
-
-    expect(lastCache.test.hasMore).toBe(true);
-  });
-
-  test('should set falsy hasMore with response', async () => {
-    const setOptionsCache = jest.fn();
-    const loadOptions = jest.fn()
-      .mockReturnValue({
-        options: [],
-      });
-
-    await requestOptions(
-      'autoload',
-      {
-        ...defaultParamsRef,
-        current: {
-          ...defaultParams,
-          loadOptions,
-          inputValue: 'test',
-        },
-      },
-      {
-        current: {},
-      },
-      0,
-      defaultSleep,
-      setOptionsCache,
-      defaultValidateResponse,
-      defaultReduceOptions,
-    );
-
-    expect(setOptionsCache).toHaveBeenCalledTimes(2);
-
-    const lastCache = setOptionsCache.mock.calls[1][0](
-      setOptionsCache.mock.calls[0][0](
-        {},
-      ),
-    );
-
-    expect(lastCache.test.hasMore).toBe(false);
   });
 });
