+++ packages/react-select-async-paginate/src/types.ts
@@ -10,6 +10,8 @@ import type {
   Props as SelectProps,
 } from 'react-select';
 
+export type RequestOptionsCallerType = 'autoload' | 'menu-toggle' | 'input-change' | 'menu-scroll';
+
 export type ReduceOptions<OptionType, Group extends GroupBase<OptionType>, Additional> = (
   prevOptions: OptionsOrGroups<OptionType, Group>,
   loadedOptions: OptionsOrGroups<OptionType, Group>,
