+++ packages/react-select-async-paginate/src/useComponents.ts
@@ -12,23 +12,11 @@ import { wrapMenuList } from './wrapMenuList';
 
 export const MenuList = wrapMenuList(defaultComponents.MenuList);
 
-type SelectComponentsConfig<
+export type SelectComponentsConfig<
 OptionType,
 IsMulti extends boolean,
 Group extends GroupBase<OptionType>,
 > = Partial<SelectProps<OptionType, IsMulti, Group>['components']>;
-
-export const useComponentsPure = <
-OptionType,
-Group extends GroupBase<OptionType>,
-IsMulti extends boolean,
->(
-    useMemoParam: typeof useMemo,
-    components: SelectComponentsConfig<OptionType, IsMulti, Group>,
-  ): SelectComponentsConfig<OptionType, IsMulti, Group> => useMemoParam(() => ({
-    MenuList,
-    ...components,
-  } as SelectComponentsConfig<OptionType, IsMulti, Group>), [components]);
 
 export const useComponents = <
 OptionType,
@@ -36,7 +24,7 @@ Group extends GroupBase<OptionType>,
 IsMulti extends boolean,
 >(
     components: SelectComponentsConfig<OptionType, IsMulti, Group>,
-  ): SelectComponentsConfig<OptionType, IsMulti, Group> => useComponentsPure(
-    useMemo,
-    components,
-  );
+  ): SelectComponentsConfig<OptionType, IsMulti, Group> => useMemo(() => ({
+    MenuList,
+    ...components,
+  } as SelectComponentsConfig<OptionType, IsMulti, Group>), [components]);
