+++ packages/react-select-async-paginate/src/wrapMenuList.tsx
@@ -16,19 +16,13 @@ import type {
 
 export const CHECK_TIMEOUT = 300;
 
-export type Props = {
+export type WrappedMenuListProps = {
   selectProps: {
     handleScrolledToBottom?: () => void;
     shouldLoadMore: ShouldLoadMore;
   };
 
   innerRef: Ref<HTMLElement>;
-
-  useEffect?: typeof useEffect;
-  useRef?: typeof useRef;
-  useCallback?: typeof useCallback;
-  setTimeout?: typeof setTimeout;
-  clearTimeout?: typeof clearTimeout;
 
   [key: string]: any;
 };
@@ -38,27 +32,20 @@ type ComponentProps = {
 };
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
-export const wrapMenuList = (MenuList: ComponentType<ComponentProps>): FC<Props> => {
-  function WrappedMenuList(props: Props) {
+export const wrapMenuList = (MenuList: ComponentType<ComponentProps>): FC<WrappedMenuListProps> => {
+  function WrappedMenuList(props: WrappedMenuListProps) {
     const {
       selectProps: {
         handleScrolledToBottom,
         shouldLoadMore,
       },
       innerRef,
-
-      useEffect: useEffectProp,
-      useRef: useRefProp,
-      useCallback: useCallbackProp,
-
-      setTimeout: setTimeoutProp,
-      clearTimeout: clearTimeoutProp,
     } = props;
 
-    const checkTimeoutRef = useRefProp(null);
-    const menuListRef = useRefProp<HTMLElement>(null);
+    const checkTimeoutRef = useRef(null);
+    const menuListRef = useRef<HTMLElement>(null);
 
-    const shouldHandle = useCallbackProp(() => {
+    const shouldHandle = useCallback(() => {
       const el = menuListRef.current;
 
       // menu not rendered
@@ -75,7 +62,7 @@ export const wrapMenuList = (MenuList: ComponentType<ComponentProps>): FC<Props>
       return shouldLoadMore(scrollHeight, clientHeight, scrollTop);
     }, [shouldLoadMore]);
 
-    const checkAndHandle = useCallbackProp(() => {
+    const checkAndHandle = useCallback(() => {
       if (shouldHandle()) {
         if (handleScrolledToBottom) {
           handleScrolledToBottom();
@@ -83,18 +70,18 @@ export const wrapMenuList = (MenuList: ComponentType<ComponentProps>): FC<Props>
       }
     }, [shouldHandle, handleScrolledToBottom]);
 
-    const setCheckAndHandleTimeout = useCallbackProp(() => {
+    const setCheckAndHandleTimeout = useCallback(() => {
       checkAndHandle();
 
-      checkTimeoutRef.current = setTimeoutProp(setCheckAndHandleTimeout, CHECK_TIMEOUT);
+      checkTimeoutRef.current = setTimeout(setCheckAndHandleTimeout, CHECK_TIMEOUT);
     }, [checkAndHandle]);
 
-    useEffectProp(() => {
+    useEffect(() => {
       setCheckAndHandleTimeout();
 
       return (): void => {
         if (checkTimeoutRef.current) {
-          clearTimeoutProp(checkTimeoutRef.current);
+          clearTimeout(checkTimeoutRef.current);
         }
       };
     }, []);
@@ -106,14 +93,6 @@ export const wrapMenuList = (MenuList: ComponentType<ComponentProps>): FC<Props>
       />
     );
   }
-
-  WrappedMenuList.defaultProps = {
-    useEffect,
-    useRef,
-    useCallback,
-    setTimeout,
-    clearTimeout,
-  };
 
   return WrappedMenuList;
 };
